% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_attrs.R
\name{all_attrs}
\alias{all_attrs}
\title{enumerate XML attributes for GENESET entities in MSigDb}
\usage{
all_attrs()
}
\description{
enumerate XML attributes for GENESET entities in MSigDb
}
\examples{
head(all_attrs())
}
