% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2docset.R
\name{cat2docset}
\alias{cat2docset}
\title{define an ssrch DocSet instance using an MSigDb category}
\usage{
cat2docset(xmlpath, category = "C7", start = 1, end = 10, ...)
}
\arguments{
\item{xmlpath}{character(1) path to full XML}

\item{category}{character(1) name of category to use}

\item{\dots}{passed to ssrch::parseDoc}
}
\description{
define an ssrch DocSet instance using an MSigDb category
}
\note{
assumes full MSigDb XML document present on disk
}
