% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_tools.R
\name{geneset_categories}
\alias{geneset_categories}
\title{enumerate MSigDb GENESET categories}
\usage{
geneset_categories()
}
\description{
enumerate MSigDb GENESET categories
}
\examples{
head(geneset_categories())
}
