% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmark}
\alias{hallmark}
\title{first draft of DocSet for Hallmark gene sets}
\format{ssrch DocSet instance}
\source{
MSigDb 6.2 xml, see inst/scripts for method
}
\usage{
hallmark
}
\description{
first draft of DocSet for Hallmark gene sets
}
\examples{
hallmark
ssrch::searchDocs("BRCA2", hallmark)
}
\keyword{datasets}
