% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immu300}
\alias{immu300}
\title{first draft of DocSet for 300 sets from C7 (immunologic)}
\format{ssrch DocSet instance}
\source{
MSigDb 6.2 xml, see inst/scripts for method
}
\usage{
immu300
}
\description{
first draft of DocSet for 300 sets from C7 (immunologic)
}
\examples{
immu300
}
\keyword{datasets}
